Renamed service back to user-manager Changed User id generation to IDENTITY Accordingly changed liquibase changelog
