+++ user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/UserManagerApplication.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.userauthservice;
+package com.crmfoodestablishment.usermanager;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class UserAuthServiceApplication {
+public class UserManagerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(UserAuthServiceApplication.class, args);
+		SpringApplication.run(UserManagerApplication.class, args);
 	}
 
 }
