+++ user-manager-service/src/main/java/com/crmfoodestablishment/usermanager/crud/entity/Role.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.userauthservice.usermanager.entity;
+package com.crmfoodestablishment.usermanager.crud.entity;
 
 public enum Role {
     CLIENT,
