+++ user-manager-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=user-auth-service
+spring.application.name=user-manager
 server.port=0
 
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
