+++ user-manager-service/src/test/java/com/crmfoodestablishment/usermanager/UserManagerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.userauthservice;
+package com.crmfoodestablishment.usermanager;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
